+++ CHANGELOG.md
@@ -9,7 +9,6 @@ Changes that are planned but not implemented yet:
 
 * Add ability for an instruction to have order agnostic operands. That is, `instr a, b` is the same as `instr b, a`. This allows `swap` to occupy less instruction space.
 * Enable instruction aliases that allow alternative mnemonics for an instruction. For example, allowing `call` and `jsr` to mean the same thing.
-* Create a memory block macro that allows the definition of memory blocks at specific addresses without altering the current address for compilation. This would be used for creatin RAM variables from ROM area code. Also used to place a specific block of code at a specific address. Different from `.org` in that there is an end of the block and at that point the address contniues on from what it was before the start of the block.
 * Normalize use of `byte_code` or `bytecode` in instruction set configuration file
 * Improve error checking:
   * Disallowed operands
@@ -22,7 +21,8 @@ Changes that are planned but not implemented yet:
 * Fixed bug where zero operand macros did not get recognized correctly
 * Added ability to reverse the order of the byte code that an instruction's operands emit.
 * Added the `relative_address` operand type.
-* Added more examples 
+* Added more examples
+* Added named memory zones feature
 
 ## [0.2.3]
 * Made instructions case insensitive